This is really two changes in one but given interdependencies and expected review latency I needed to combine the CLs.
1. Made the 5* toolpath build using the go build
   system. Hooked the subdirectories to clean.bash but added a
   separate make5.bash for now. Minor massage to make the code
   more similar to the current structure of 6c/6a/6l.

2. Change all references from long to int32 in line with
   similar change for the other toolchains.

The end result is that 5c, 5a and 5l can now be compiled and
the executables start up properly. Haven't thrown any input at
them yet.

R=rsc
APPROVED=rsc
DELTA=1052  (392 added, 328 deleted, 332 changed)
OCL=26757
CL=26761

+++ src/cmd/5c/sgen.c
@@ -92,7 +92,7 @@ codgen(Node *n, Node *nn)
 void
 supgen(Node *n)
 {
-	long spc;
+	int32 spc;
 	Prog *sp;
 
 	if(n == Z)
@@ -113,7 +113,7 @@ gen(Node *n)
 	Node *l, nod;
 	Prog *sp, *spc, *spb;
 	Case *cn;
-	long sbc, scc;
+	int32 sbc, scc;
 	int o, f;
 
 loop:
